% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_input.R
\name{format_input}
\alias{format_input}
\title{Pre-Train Data Formatting}
\usage{
format_input(x, y, data)
}
\arguments{
\item{x}{A \code{character vector} containing the names of all features
we will train the binary classification on.}

\item{y}{A \code{character} containing the name of the classified feature
(which we will fit the model on).}

\item{data}{A \code{data frame} containing the data we will train the binary
classification on.}
}
\value{
A \code{list} containing the following attributes, where n is the number
of observations in the given data frame and k is the number of provided features
to predict on:
\describe{
     \item{x}{An n x k+1 matrix containing the predictors we will use to train}
     \item{y}{An n x 1 vector contatining the classification of each observation}
}
}
\description{
Format the data correctly so that it can be trained,
including changing the format of factors.
}
\examples{
format_input(x = c("age", "workclass", "house.per.week"),
      y = "sex", data = adult)
}
\author{
Emily Knight
}
