% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_function.R
\name{loss_function}
\alias{loss_function}
\title{Loss Function}
\usage{
loss_function(beta, x, y)
}
\arguments{
\item{beta}{A \code{vector} containing the beta that we will evaluate the
loss function for.}

\item{x}{A \code{matrix} containing the formatted feature data.}

\item{y}{A \code{vector} containing the classifications of the features.}
}
\value{
A \code{numeric} which is the log likelihood value.
}
\description{
Define the loss function that we will use to optimize for
binary classification.
}
\author{
Emily Knight
}
